Update front matter

Updates the front matter of a Markdown file on disk.

Usage

update_frontmatter(
  md_file_path,
  rmd_file,
  title = NULL,
  sidebar_label = NULL,
  sidebar_order = NULL
)

Arguments

md_file_path

Path to the Markdown file on disk.

rmd_file

Path to the R Markdown file, either a local path or a URL.

title

Title of the article, to show on top of the page.

sidebar_label

Title in the sidebar.

sidebar_order

Number indicating the order of the article in the sidebar.

Value

Markdown file with updated front matter, written to disk.

Examples

if (FALSE) { # \dontrun{
update_frontmatter(
  md_file_path = file.path(
    "output/src/content/docs/software/gcube/occurrence-process.md"
  ),
  rmd_file = file.path(
    "https://raw.githubusercontent.com/b-cubed-eu/gcube/refs/heads/main",
    "vignettes/articles/occurrence-process.Rmd"
  ),
  title = "2. Occurrence process",
  sidebar_label = "Occurrence process",
  sidebar_order = 2
)
} # }